Day2

Documentation

data Color

Constructors

Red

Green

Blue

type Cube = (Int, Color)

type GroupOfCubes = (Int, Int, Int)

Group of cubes for part two, representing (Red, Green Blue)

type SetOfCubes = [Cube]

A set of cubes is a tuple of the number of red, green and blue cubes

type Game = (Int, [SetOfCubes])

A game is a tuple of the id of the game a a list of set of cubes

isValidGame :: [SetOfCubes] -> Bool

isValidSetOfCubes :: SetOfCubes -> Bool

isValidCube :: Cube -> Bool

extractValidGameId :: Game -> Int

sumValidGames :: [Game] -> Int

powerOfGame :: Game -> Int

multiplyGroupOfCubes :: GroupOfCubes -> Int

minNumberOfCubesOnGame :: [SetOfCubes] -> GroupOfCubes

minGroupOfCubes :: GroupOfCubes -> GroupOfCubes -> GroupOfCubes

minNumberOfSetOfCubes :: SetOfCubes -> GroupOfCubes

minNumberOfCube :: Cube -> GroupOfCubes -> GroupOfCubes

sumPowerOfGames :: [Game] -> Int

parseGame :: String -> Game

parseGameId :: String -> Int

extractGameId :: String -> Int

parseGameInfo :: String -> [SetOfCubes]

parseGameSets :: String -> [SetOfCubes]

parseSetOfCubes :: String -> SetOfCubes

extractNumberOfCubes :: String -> Cube

parseCubeColor :: String -> Color

executeDay2Part1 :: IO ()

executeDay2Part2 :: IO ()