Y2015.D09

Description

Solutions to the day 09 set of problems for adventofcode.com.

Documentation

longestRoute

Arguments

:: [Route]
List of route flight paths

-> Maybe Distance
Possibly longest distance

Finds the longest route given a list of routes

routeParser

Arguments

:: Parser [Route]
Route parser

Parsec parser for the Route type

shortestRoute

Arguments

:: [Route]
List of route flight paths

-> Maybe Distance
Possibly shortest distance

Finds the shortest route given a list of routes