Y2015.D17

Description

Solutions to the day 17 set of problems for adventofcode.com.

Documentation

filledAmong

Arguments

:: Int
Length

-> String
List of container sizes

-> Int
Number of combinations

Finds how many combinations of containers can fit a volume of nog

minFilledAmong

Arguments

:: Int
Length

-> String
List of container sizes

-> Int
Number of combinations

Finds how many combinations of containers the ideal combination can |fit a volume of nog