Y2015.D24

Description

Solutions to the day 24 set of problems for adventofcode.com.

Documentation

idealEntanglement

Arguments

:: Int
How many compartments to divide amongst.

-> String
Input as a string containing newlined-separated list of package weights.

-> Int
Ideal entanglement value for the resulting package distribution.

Find the ideal entanglement value for a given input of packages.

idealEntanglementOptimized

Arguments

:: Int
How many compartments to divide amongst.

-> String
Input as a string containing newlined-separated list of package weights.

-> Maybe Int
Ideal entanglement value for the resulting package distribution if one can be found.

Find the ideal entanglement value for a given input of packages.