Y2021

Description

Solutions to the set of problems for adventofcode.com.

Each day is broken up into an individual module with accompaying spec tests and (possibly?) benchmarks. Each day's module exports are re-exported here for convenience when importing Y2021 et. al.

This year I leaned a little bit more into making things very Haskell-ly by relying more on laziness, centralizing solutions around data types, etc.

Documentation

module Y2021.D09

module Y2021.D08

module Y2021.D07

module Y2021.D06

module Y2021.D05

module Y2021.D04

module Y2021.D03

module Y2021.D02

module Y2021.D01