Y2021.D05

Description

Solutions to the 2021 day 05 set of problems for adventofcode.com.

Documentation

type Point = (Int, Int)

Makes some signatures easier to read

type Ray = [(Point, Point)]

Makes some signatures easier to read

part5A :: Text -> Int

Solve part A

part5AHM :: Text -> Int

Solve part A, but with a HashMap

part5B :: Text -> Int

Solve part B

part5BHM :: Text -> Int

Solve part B, but with a HashMap

solve5 :: (Ray -> Ray) -> Text -> Int

Higher-order function solution to share parts A and B.

solve5HM :: (Ray -> Ray) -> Text -> Int

Same as solve5, but using HashMap.

lineTo :: Point -> Point -> [(Point, Sum Int)]

Accept a start and end point and return a list of points that draw a line to the endpoint. Note that this doesn't work for anything other than vertical, horizontal, and 45deg.

(~~) :: Int -> Int -> [Int] infixl 5

Messing around with a custom operator for ranges that can handle both up-to and down-to.

parseVents :: Text -> Ray

Parse puzzle input into simple pairs of pairs of points.