Y2021.D06

Description

Solutions to the 2021 day 06 set of problems for adventofcode.com.

Documentation

part6A :: Text -> Int

Solve part A

part6AMV :: Text -> Int

Solve part A, with unboxed mutable vectors

part6B :: Text -> Int

Solve part B

part6BMV :: Text -> Int

Solve part B, with unboxed mutable vectors

solve6MV :: Int -> [Int] -> Int

Solution algorithm using super aggressive mutable unboxed vectors.

solve6 :: Int -> [Int] -> Int

Iterate for a given number of days given a starting value and return the total.

breed :: Num a => IntMap a -> IntMap a

Simulate one day passing for the collection of fish.

parseFish :: Text -> [Int]