Y2021.D08

Description

Solutions to the 2021 day 08 set of problems for adventofcode.com.

Documentation

data Signal

Guess I'll use a sum type for this

Constructors

SigA

SigB

SigC

SigD

SigE

SigF

SigG

type SignalSequence = [Signal]

These are just for readability

type SignalSet = Set Signal

type SignalEntries = [([SignalSequence], [SignalSequence])]

data Display

Another sum type for "display"

Constructors

D0

D1

D2

D3

D4

D5

D6

D7

D8

D9

part8A :: Text -> Int

Solve part A

solve8A :: [(a, [SignalSequence])] -> Int

This one is for benching apart from the parsing overhead.

part8APredicate :: SignalSequence -> Bool

Our simple predicate for part A per the instructions.

part8B :: Text -> Int

Solve part B

decode :: [SignalSequence] -> [SignalSequence] -> Sum Int

Brute-force the input to decode the output.

untangle :: [SignalSequence] -> Maybe (HashMap Signal Signal)

correctSet :: Set SignalSet

Compile the mappings for the golden input.

sigMap :: HashMap SignalSet Display

Mappings from the right signals to their associated displays.

dMap' :: Display -> SignalSequence

From Display -> the right sequence.

parseSevSeg :: Text -> SignalEntries

Parse puzzle input into a list of Ints with faster attoparsec.