Y2023.D01

Description

Solutions to the 2023 day 01 set of problems for adventofcode.com.

Documentation

part1A :: String -> Int

Solve for part A

part1B :: String -> Int

Solve for part B

base10Tuple :: (Int, Int) -> Int

Turn a tuple into a two-digit number

intoPair :: [Int] -> Maybe (Int, Int)

Given a list of ints, try and turn it into a tuple of first and last elements, with Nothing as the result if it can’t be done.

calibrateA :: String -> [[Int]]

Transform an input string into a list of all the digits found in each line.

calibrateB :: String -> [[Int]]

Transform an input string into a list of all the digits and spelled numbers found in each line.