Y2015.D05

Description

Solutions to the day 05 set of problems for adventofcode.com.

Documentation

clean

Arguments

:: String
Input string.

-> Bool
Whether the string is clean.

Predicate to determine whether a string eschews forbidden strings.

isNice

Arguments

:: String
Test input string.

-> Bool
Whether the given input string is nice.

Predicate to determine whether a given string is "nice".

isNicer

Arguments

:: String
Test input string

-> Bool
Whether string is nice

Whether a given string is "nice" per the spec.

thriceVoweled

Arguments

:: String
Input string to test.

-> Bool
Whether the string passes the predicate.

Predicate to determine whether the given string contains at least three |vowels.

twiceRow

Arguments

:: String
Input string to test.

-> Bool
Whether the given string passes the predicate.

Predicate to determine whether a given string contains two letters |in a row.