Y2021.D04

Description

Solutions to the 2021 day 04 set of problems for adventofcode.com.

Documentation

data Bingo a

Fancy data type to represent _game_ state, not just a card

Constructors

Ended (Int, a)

Playing a

type Card = [[Square]]

Represents a game board state

data Square

Small wrapper over how to record marked/unmarked squares

Constructors

Marked

Unmarked Int

part4A :: Text -> Int

Solve part A

part4B :: Text -> Int

Solve part B

solve4 :: ([Bingo Card] -> Bingo Card) -> [Int] -> [Bingo Card] -> Int

Fortunately both A and B are just asking slightly different questions, so we have a higher-order function to determine how to pull the matching value from our resultant list.

tally :: Bingo Card -> Int

Problem-defined method of scoring a card. The "in-progress" scoring isn't actually ever used. Probably a bug.

sumCard :: Card -> Int

Yank out the total values for a card.

iterateMap :: [Int] -> [Bingo Card] -> [Bingo Card]

Our main recursive loop to walk through the "called" numbers. We return a list solely of elements that meet the "ended" predicate.

mark :: Int -> [Bingo Card] -> [Bingo Card]

Accept a list of `Bingo Card`s and update each with the called number.

markRows :: Int -> Bingo Card -> Bingo Card

Update a Card - we make the determination to "end" a game here, so as soon as its "won" the type swaps over to the ended value.

hasEnded :: Bingo a -> Bool

Small helper predicate to tell is when a Bingo type is over or not.

gameEnd :: Int -> Bingo Card -> Bingo Card

Important function to determine if a card is a winning card. If so it changes its type.

marked :: Square -> Bool

Another little helper to tell us if a spot has been marked or not.

bingoParser :: Parser ([Int], [Bingo Card])

This is a sort of hairy, but all-in-one, parser for the problem set input. 5 is a magic number for board size.