Y2021.D08

Description

Solutions to the 2021 day 08 set of problems for adventofcode.com.

Documentation

part8A :: Text -> Int

Solve part A

part8B :: Text -> Int

Solve part B

parse8 :: Text -> SignalEntries

Parse puzzle input into a list of Ints with faster attoparsec.

solve8A :: [(a, [SignalSequence])] -> Int

This one is for benching apart from the parsing overhead.